new header
header with images and learned to position images in the background using background-size: cover and background-position: wherever the image needs to be positioned. I am working on adding some links to other pages within the site that are aesthetically pleasant. I am also trying to make the header with these images just look nicer.